# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Center API",
    "summary": "Документ-центр ИП «Северный Контур»: проекты, документы, версии и аудит",
    "description": "\nВнутренний REST API ИП «Северный Контур» для управления документами по проектам.\n\n**Базовый путь публичного контракта:** `/api/v1`. Версия контракта (`v1`)\nне совпадает с версией приложения, указанной в заголовке документации.\n\n### Аутентификация\nПолучите токен операцией `POST /api/v1/auth/login` и передавайте его в заголовке\n`Authorization: Bearer <token>`. В Swagger UI нажмите **Authorize** и вставьте токен.\n\n### Роли\n- **admin** — полный доступ ко всем проектам, пользователям и журналу действий;\n- **manager** — создаёт проекты и выдаёт доступ участникам своих проектов;\n- **worker** — работает с документами в рамках выданного доступа `editor`;\n- **viewer** — только чтение в рамках выданного доступа.\n\n### Ошибки\nВсе ошибки возвращаются в формате RFC 9457 Problem Details\n(`application/problem+json`) с полями `type`, `title`, `status`, `detail`,\n`instance`; ошибки валидации дополнительно содержат массив `errors`.\n\n### Коллекции\nКоллекции поддерживают параметры `offset` (≥ 0) и `limit` (1…100, по умолчанию 20)\nи возвращают конверт `items`, `total`, `offset`, `limit`.\n",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Получить список проектов",
        "description": "Вернуть проекты, доступные текущему пользователю.\n\nАдминистратор получает все проекты. Остальные роли — проекты,\nкоторыми владеют, и проекты с выданным доступом.",
        "operationId": "list_projects",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение от начала коллекции",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение от начала коллекции"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы, максимум 100",
              "default": 20,
              "title": "Limit"
            },
            "description": "Размер страницы, максимум 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Страница проектов, доступных пользователю",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectRead_"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Создать проект",
        "description": "Создать проект и назначить текущего пользователя его владельцем.\n\nОперация доступна ролям admin и manager.",
        "operationId": "create_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Созданный проект",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Получить проект",
        "description": "Вернуть проект, доступный текущему пользователю.\n\nНедоступный проект представляется как несуществующий: факт наличия\nчужого проекта не раскрывается.",
        "operationId": "get_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "responses": {
          "200": {
            "description": "Карточка проекта",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Изменить проект",
        "description": "Изменить название или описание проекта.\n\nОперация доступна владельцу проекта и администратору. Неуказанные\nполя не изменяются.",
        "operationId": "update_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Изменённый проект",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Удалить проект",
        "description": "Удалить проект вместе с зависимыми записями.\n\nОперация доступна владельцу проекта и администратору.",
        "operationId": "delete_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "responses": {
          "204": {
            "description": "Проект удалён"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Путь к полю запроса",
            "examples": [
              "title"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Причина отклонения значения",
            "examples": [
              "String should have at least 3 characters"
            ]
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "description": "Описание одной ошибки валидации поля."
      },
...
      "Page_ProjectRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectRead"
            },
            "type": "array",
            "title": "Items",
            "description": "Записи текущей страницы"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Общее количество записей, доступных запросу"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Смещение текущей страницы от начала коллекции"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Запрошенный размер страницы"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page[ProjectRead]",
        "description": "Страница коллекции с метаданными пагинации: items, total, offset, limit"
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный URI типа ошибки"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Краткое название типа ошибки, не зависящее от экземпляра"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "description": "HTTP-код ответа"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Описание конкретного случая возникновения ошибки"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "Путь запроса, при обработке которого возникла ошибка"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ProblemDetails",
        "description": "Тело ответа с ошибкой по RFC 9457.",
        "examples": [
          {
            "detail": "Project 42 does not exist or is unavailable",
            "instance": "/api/v1/projects/42",
            "status": 404,
            "title": "Project not found",
            "type": "https://docs.local/problems/project-not-found"
          }
        ]
      },
...
      "ProjectCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Title",
            "description": "Название проекта, от 3 до 120 символов",
            "examples": [
              "Реконструкция котельной"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Описание проекта"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ProjectCreate",
        "description": "Данные для создания проекта."
      },
      "ProjectRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Идентификатор проекта"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Название проекта"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Описание проекта"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id",
            "description": "Идентификатор владельца проекта"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Момент создания проекта",
            "examples": [
              "2026-09-09T08:15:30Z"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "owner_id",
          "created_at"
        ],
        "title": "ProjectRead",
        "description": "Карточка проекта."
      },
      "ProjectUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Новое название проекта"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Новое описание проекта"
          }
        },
        "type": "object",
        "title": "ProjectUpdate",
        "description": "Изменяемые атрибуты проекта; неуказанные поля не изменяются."
      },
...
      "ValidationProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный URI типа ошибки"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Краткое название типа ошибки, не зависящее от экземпляра"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "description": "HTTP-код ответа"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Описание конкретного случая возникновения ошибки"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "Путь запроса, при обработке которого возникла ошибка"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors",
            "description": "Перечень полей, не прошедших валидацию"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ValidationProblemDetails",
        "description": "Тело ответа для ошибок валидации запроса (HTTP 422).\n\nРасширяет базовую структуру членом ``errors`` согласно STD-ERR-05.",
        "examples": [
          {
            "detail": "Request body does not match the expected schema",
            "errors": [
              {
                "field": "title",
                "message": "String should have at least 3 characters"
              }
            ],
            "instance": "/api/v1/projects",
            "status": 422,
            "title": "Request validation failed",
            "type": "https://docs.local/problems/validation-error"
          }
        ]
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}